+++ include/setup.h
@@ -33,12 +33,13 @@ enum ReturnCodes
 struct global_t {
     bool isVersion;            		// -V option, flag
     bool isCounterElectrode;   		// -c option, flag
-    bool isVerbose;          		// -v option, flag
+    int verbosity;          		// -v option, flag
     bool isTimer;					// -b option, flag
     std::string listFilename;		// -f option, list file name
 	std::string outputFilename;		// -o option, output file name
     std::vector<std::string> fileNames;	// list of filenames
 	std::map<std::string, std::string> device;	// map of file ext and instrument desc
+	std::string globalDevice;       // device for global table
 	std::map<ReturnCodes, std::string> plots;	// map of experimental data types
     size_t fileSizeSum;		       	// total file size of all input files
     size_t recordsSum;       		// total data records